Python: Add up-/downgrade scripts
Adds up- and downgrade scripts for the support for type parameter
defaults. Since one direction adds no data, and the other simply removes
the existing data, I left the scripts empty.

Based on the
[guide](https://github.com/github/codeql/blob/main/docs/prepare-db-upgrade.md),
I've put the compatibility as

full: results from the upgraded snapshot will be identical to results
      from a snapshot built with the new version of the toolchain.

In the upgrade direction this should be true. If a database was
successfully built with the old toolchain, then it doesn't contain any
type parameter defaults, nor will it after the upgrade.

In the downgrade direction, the snapshot _might_ contain type parameter
defaults, in which case these will be removed when downgrading. But in
that case, it's not possibly to build a snapshot with the target
toolchain (the old one), so the compatibility holds vacuously.